refactor(web): extract generic layout hooks into components/hooks.ts (issue #642, Phase 3 seam 4 pt 1)
useSlideSize, useMediaQuery (+ its MOBILE_MEDIA_QUERY/PORTRAIT_MEDIA_QUERY
constants), useIsMobile, useFullscreen, and useThumbColumnSize were plain
module-local hooks in App.tsx with no dependency on any App.tsx-private
state; they measure DOM elements and viewport media queries generically,
so they move to components/hooks.ts alongside the existing Theme/
ThemeContext content. Verbatim move (export added on the four hooks
App.tsx still calls; useMediaQuery and the two media-query constants stay
module-private in the new file, since nothing outside useIsMobile used
them) — no behavior change.
